+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/embedded/content/EmbeddedContent.kt
@@ -5,6 +5,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Immutable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.stripe.android.paymentelement.ExperimentalEmbeddedPaymentElementApi
@@ -18,10 +19,30 @@ import com.stripe.android.uicore.StripeTheme
 internal data class EmbeddedContent(
     private val interactor: PaymentMethodVerticalLayoutInteractor,
     private val embeddedViewDisplaysMandateText: Boolean,
-    private val rowStyle: Embedded.RowStyle
+    private val rowStyle: Embedded.RowStyle,
+    private val isImmediateAction: Boolean,
 ) {
     @Composable
     fun Content() {
+        /**
+         * This validation is here because of a weird interaction with 2-Step Integration.
+         *
+         * If this were in configure or when state is set, then it would fail for the 1st instance of embedded
+         * in the 2 step integration because a user would not set a rowSelectionBehavior on the 1st instance of embedded
+         * because the 1st instance doesn't show a UI. However, the first instance still has to be configured so it will
+         * fail unless the user sets an empty rowSelection ImmediateAction callback.
+         *
+         * Having validation here ensures that we only validate when the embedded content is shown.
+         */
+        LaunchedEffect(rowStyle, isImmediateAction) {
+            if (rowStyle is Embedded.RowStyle.FlatWithChevron && !isImmediateAction) {
+                throw IllegalArgumentException(
+                    "EmbeddedPaymentElement.Builder.rowSelectionBehavior() must be set to ImmediateAction when using " +
+                        "FlatWithChevron RowStyle. Use a different style or enable ImmediateAction rowSelectionBehavior"
+                )
+            }
+        }
+
         StripeTheme {
             Column(
                 modifier = Modifier